requirements.txt
accelerate==1.10.0
aiofiles==24.1.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
backoff==2.2.1
bcrypt==4.3.0
beanie==2.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.3.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==45.0.6
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dataclasses-json==0.6.7
decorator==4.4.2
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
effdet==0.4.1
emoji==2.14.1
et_xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.59.1
fsspec==2025.5.1
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.178.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-vision==3.10.2
google-genai==1.30.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
groq==0.29.0
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
hf-xet==1.1.7
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.8
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.9
langdetect==1.0.9
lazy-model==0.3.0
llvmlite==0.44.0
lxml==6.0.0
Markdown==3.8.2
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
more-itertools==10.7.0
motor==3.7.1
moviepy==1.0.3
mpmath==1.3.0
msoffcrypto-tool==5.4.2
mypy_extensions==1.1.0
networkx==3.4.2
nltk==3.9.1
numba==0.61.2
numpy==1.26.4
olefile==0.47
omegaconf==2.3.0
onnx==1.18.0
onnxruntime==1.22.1
openai-whisper==20250625
opencv-python==4.12.0.88
openpyxl==3.1.5
packaging==25.0
pandas==2.3.1
passlib==1.7.4
pdf2image==1.17.0
pdfminer.six==20250506
pi_heif==1.1.0
pikepdf==9.10.2
pillow==11.3.0
proglog==0.1.12
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.10
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.14.0
pypandoc==1.15
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-iso639==2025.2.18
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.6.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
SoundCard==0.4.4
sounddevice==0.5.2
soupsieve==2.7
starlette==0.47.2
stripe==12.4.0
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.9.0
timm==1.0.19
tokenizers==0.21.4
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
transformers==4.55.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
unstructured==0.18.13
unstructured-client==0.42.3
unstructured-inference==1.0.5
unstructured.pytesseract==0.3.15
uritemplate==4.2.0
urllib3==2.2.2
uvicorn==0.35.0
uvloop==0.21.0
waitress==3.0.2
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xlrd==2.0.2
xlsxwriter==3.2.5
